Refactor tracking of in use and free pids
We now track all members across pools in the #state.all_members dict.
This simplifies the handling of returned members and dealing with
crashes of members that are not currently in use.

The all_members dict maps member pid to a tuple of {PoolName,
ConsumerPid | free, Time} where Time is a timestamp that gets updated
each time the member is returned to the pool.  This sets things up to
cull unused members after a traffic spike.