Added use_member/1 to reduce code noise
In many cases the process pool will follow the standard pattern of
Get->Use->Return, and hence there'll be a lot of code scattered
through applications which follows this pattern. Rather than leave
the responsibility of following this pattern to the caller, the
use_member/1 function was added to cater for this primary case.

This function catches exceptions and returns the member to the pool
with an indication of error so that the process can be restarted.
If the function succeeds then the process is returned untouched.

Added some documentation to the README so that it's a bit more
visible to future users.