Limit attempts to add members while serving take_member request
This patch introduces a new pool configuration configuration parameter
'add_member_retry' with a default value of 1. This is the number of
times pooler will attempt to add a new member to the pool while
servicing a take_member request if there are no free members in the
pool and the member count is less than max_count.

Previously pooler would loop forever trying to add a new member.  This
could cause take_member not to return quickly if pooler was unable to
add any members. With this patch, error_no_members is returned after
add_member_retry attempts have been made to add a new member to the
pool.