Fix do_return_member to properly handle an unknown pid
The test for this was passing because we were only testing the client
side. pooler was actually crashing and restarting and we didn't test
for that. Now that return member is a call instead of a cast, the
server crash induces a client crash and this bug is easy to see.

This fix is simply to ignore unknown Pids in do_return_member as we
were already doing for status fail.