Add load balancing support using pg2 and an ets table
Pool configs with a group value other than 'undefined' are added to a
pg2 process group at pool startup. The pool joins the group when it
receives a timeout message triggered by init. This is done in order to
ensure that the pool is ready to serve requests as soon as it can be
seen as a member of a group.

In order to provide a simple API where only the group name is needed
to take and return a member, members that are access via the new
take_group_member/1 function are added to an ETS table that maps the
member pid to the pid of the pool to which the member belongs. When
returning a member via return_group_member/3 the ETS table is used to
locate the pid of the appropriate pool and return_member is then
called. At present the group name is not actually used upon return,
but keeping the symmetry seems reasonable and will allow more options
for some alternative implementations later if desired.

Whenever a member is returned, the code calls do_return_member and
this function now ensure that a call to clean_group_table is
made. Only when the pool's group value is other than 'undefined' this
function attempts to delete the entry for the member pid in the ETS
table that is used for tracking groups.

The pooler_sup is responsible for starting the ETS table. If a pool
crashes, it could orphan entries in the ETS table. The pg2 group will
remove processes that die.