Use metrics_mod stored in pool record in send_metric
The metrics module specified in app config is now stored as part of
the pool record. This allows send_metric in pooler.erl to determine
whether or not metrics are enabled by using data in the server state
rather than having to retrieve data from app config for each metric
send call. The only downside is duplicating the metrics module name
across all pools.

The default value is pooler_no_metric and will make calls to
send_metric a no-op.