Set nowarn_deprecated_type erl_opt for R17 compat
R17 deprecates built-in type spec for dict() in favor of
dict:dict(). However, the latter is not available in R16 and below and
is a compile time error. Ignoring the warning isn't great, but allows
the code to compile clean on R16 and R17.

Also updating the list of supported/tested Erlang versions in the travis
yaml.