Remove the grouped pool feature
The grouped pool feature has never been used by the authors of this
library and does not align with the primary uses of pooler.

The change is motivated by testing pooler on Erlang 18 and seeing
deprecation warnings for `erlang:now` which was being used for RNG
seeding.

With the removal of grouped pools, pooler compiles and tests cleanly on
Erlang 18.